-- Update Candidate endpoint added -- Delete Candidate endpoint added.

+++ KHRMS.Infrastructure/Repositories/GenericRepository.cs
@@ -12,7 +12,7 @@ namespace KHRMS.Infrastructure
             _dbContext = context;
         }
 
-        public async Task<T> GetById(int id)
+        public async Task<T> GetById(long id)
         {
             return await _dbContext.Set<T>().FindAsync(id);
         }
